refactor: FileUtils를 Utils로 변경하고 getHttpVersion 메소드를 이동
